+++ src/index.ts
@@ -43,4 +43,4 @@ function useModuleComputed<State extends StateRecord, Actions extends ActionsRec
 
 export { defineModule, defineProvider, useScopeContext, useModule, useModuleActions, useModuleComputed };
 
-export const VERSION = '0.3.6';
+export const VERSION = '0.4.0';